refactor: mvc 패턴의 구조로 패키지 구조 변경
